requirements.txt
flask
gunicorn
torch
transformers
Pillow
requests
python-dotenv
openai
